switch data type to float

+++ FebLaunch/DFRobot_BMX160.h
@@ -783,11 +783,11 @@ struct bmx160Dev {
 //struct BMX160_sensor_data {
 struct bmx160SensorData {
   /*! X-axis sensor data */
-  int16_t x;
+  float x;
   /*! Y-axis sensor data */
-  int16_t y;
+  float y;
   /*! Z-axis sensor data */
-  int16_t z;
+  float z;
   /*! sensor time */
   uint32_t sensortime;
 };
